+++ lib/repo_analyzer.rb
@@ -1,4 +1,5 @@
 require "bundler/audit"
+require "bundler/stats"
 require "brakeman"
 require "fasterer"
 require "octokit"