+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+### Unreleased
+
+#### Added
+
+* Ruby 3.2 support.
+
 ### v1.2.0
 
 #### Added