+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Unreleased
+
+#### Added
+
+* Add extract [reek](https://github.com/troessner/reek) data.
 
 ### v1.3.0