+++ CHANGELOG.md
@@ -1,9 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
-
-### Unreleased
-
+### v1.4.0
 #### Added
 
 * Add [reek](https://github.com/troessner/reek) extractor.